+++ internal/app/repositories/token_mock.go
@@ -89,11 +89,11 @@ func (mr *MockTokenRepositoryMockRecorder) FindById(ctx, id any) *gomock.Call {
 }
 
 // List mocks base method.
-func (m *MockTokenRepository) List(ctx context.Context, limit, offset int32) ([]models.Token, int, error) {
+func (m *MockTokenRepository) List(ctx context.Context, limit, offset uint64) ([]models.Token, uint64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "List", ctx, limit, offset)
 	ret0, _ := ret[0].([]models.Token)
-	ret1, _ := ret[1].(int)
+	ret1, _ := ret[1].(uint64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
