+++ internal/app/repositories/token_test.go
@@ -51,9 +51,9 @@ func Test_TokenRepository_List(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		limit    int32
-		offset   int32
-		total    int
+		limit    uint64
+		offset   uint64
+		total    uint64
 		expected []models.Token
 	}{
 		{
