+++ internal/app/repositories/user_test.go
@@ -43,9 +43,9 @@ func Test_UserRepository_List(t *testing.T) {
 
 	tests := []struct {
 		name     string
-		limit    int32
-		offset   int32
-		total    int
+		limit    uint64
+		offset   uint64
+		total    uint64
 		expected []models.User
 	}{
 		{
