+++ internal/app/serializers/pagination.go
@@ -1,9 +1,9 @@
 package serializers
 
 type PaginationMeta struct {
-	Page  int `json:"page"`
-	Per   int `json:"per"`
-	Total int `json:"total"`
+	Page  uint64 `json:"page"`
+	Per   uint64 `json:"per"`
+	Total uint64 `json:"total"`
 }
 
 type PaginationResponse[T interface{}] struct {
