+++ internal/app/services/roles_mock.go
@@ -103,11 +103,11 @@ func (mr *MockRolesMockRecorder) FindRoleDetailsById(ctx, id any) *gomock.Call {
 }
 
 // List mocks base method.
-func (m *MockRoles) List(ctx context.Context, pagination *Pagination) ([]models.Role, int, error) {
+func (m *MockRoles) List(ctx context.Context, pagination *Pagination) ([]models.Role, uint64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "List", ctx, pagination)
 	ret0, _ := ret[0].([]models.Role)
-	ret1, _ := ret[1].(int)
+	ret1, _ := ret[1].(uint64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
