+++ internal/app/services/tokens_mock.go
@@ -88,11 +88,11 @@ func (mr *MockTokensMockRecorder) FindById(ctx, id any) *gomock.Call {
 }
 
 // List mocks base method.
-func (m *MockTokens) List(ctx context.Context, pagination *Pagination) ([]models.Token, int, error) {
+func (m *MockTokens) List(ctx context.Context, pagination *Pagination) ([]models.Token, uint64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "List", ctx, pagination)
 	ret0, _ := ret[0].([]models.Token)
-	ret1, _ := ret[1].(int)
+	ret1, _ := ret[1].(uint64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
