+++ internal/app/services/users_mock.go
@@ -118,11 +118,11 @@ func (mr *MockUsersMockRecorder) FindUserDetailsById(ctx, id any) *gomock.Call {
 }
 
 // List mocks base method.
-func (m *MockUsers) List(ctx context.Context, pagination *Pagination) ([]models.User, int, error) {
+func (m *MockUsers) List(ctx context.Context, pagination *Pagination) ([]models.User, uint64, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "List", ctx, pagination)
 	ret0, _ := ret[0].([]models.User)
-	ret1, _ := ret[1].(int)
+	ret1, _ := ret[1].(uint64)
 	ret2, _ := ret[2].(error)
 	return ret0, ret1, ret2
 }
